Extract specific collections SQL queries in traits
Similar to f5c1c6c964c3ee4672a3fecf2d384c246e77ec15.
