+++ src/models/dao/Collection.php
@@ -11,6 +11,14 @@ namespace flusio\models\dao;
 class Collection extends \Minz\DatabaseModel
 {
     use BulkQueries;
+    use LockQueries;
+    use collections\CleanerQueries;
+    use collections\DiscoveryQueries;
+    use collections\FetcherQueries;
+    use collections\OpmlImportatorQueries;
+    use collections\PocketQueries;
+    use collections\SearchQueries;
+    use collections\StatisticsQueries;
 
     /**
      * @throws \Minz\Errors\DatabaseError
@@ -19,79 +27,6 @@ class Collection extends \Minz\DatabaseModel
     {
         $properties = array_keys(\flusio\models\Collection::PROPERTIES);
         parent::__construct('collections', 'id', $properties);
-    }
-
-    /**
-     * Return the number of collections (type "collection").
-     *
-     * @return integer
-     */
-    public function countCollections()
-    {
-        $sql = <<<'SQL'
-            SELECT COUNT(*) FROM collections
-            WHERE type = 'collection'
-        SQL;
-
-        $statement = $this->query($sql);
-        return intval($statement->fetchColumn());
-    }
-
-    /**
-     * Return the number of feeds (type "feed").
-     *
-     * @return integer
-     */
-    public function countFeeds()
-    {
-        $sql = <<<'SQL'
-            SELECT COUNT(*) FROM collections
-            WHERE type = 'feed'
-        SQL;
-
-        $statement = $this->query($sql);
-        return intval($statement->fetchColumn());
-    }
-
-    /**
-     * Return the number of feeds (type "feed") by hours.
-     *
-     * @return integer[]
-     */
-    public function countFeedsByHours()
-    {
-        $sql = <<<'SQL'
-            SELECT TO_CHAR(feed_fetched_at, 'HH24') AS hour, COUNT(*) as count
-            FROM collections
-            WHERE type = 'feed'
-            GROUP BY hour
-        SQL;
-
-        $statement = $this->query($sql);
-        $result = $statement->fetchAll();
-        $count_by_hours = [];
-        foreach ($result as $row) {
-            $count_by_hours[$row['hour']] = intval($row['count']);
-        }
-        ksort($count_by_hours);
-        return $count_by_hours;
-    }
-
-    /**
-     * Return the number of collections (type "collection").
-     *
-     * @return integer
-     */
-    public function countCollectionsPublic()
-    {
-        $sql = <<<'SQL'
-            SELECT COUNT(*) FROM collections
-            WHERE type = 'collection'
-            AND is_public = true
-        SQL;
-
-        $statement = $this->query($sql);
-        return intval($statement->fetchColumn());
     }
 
     /**
@@ -114,45 +49,6 @@ class Collection extends \Minz\DatabaseModel
         $statement = $this->prepare($sql);
         $statement->execute([
             ':link_id' => $link_id,
-        ]);
-        return $statement->fetchAll();
-    }
-
-    /**
-     * Return the list of public collections attached to the given topic.
-     *
-     * @param string $topic_id
-     * @param integer $pagination_offset
-     * @param integer $pagination_limit
-     *
-     * @return array
-     */
-    public function listPublicByTopicIdWithNumberLinks($topic_id, $pagination_offset, $pagination_limit)
-    {
-        $sql = <<<'SQL'
-            SELECT c.*, COUNT(lc.*) AS number_links
-            FROM collections c, collections_to_topics ct, links_to_collections lc, links l
-
-            WHERE c.id = ct.collection_id
-            AND c.id = lc.collection_id
-            AND l.id = lc.link_id
-
-            AND c.is_public = true
-            AND l.is_hidden = false
-            AND ct.topic_id = :topic_id
-
-            GROUP BY c.id
-
-            ORDER BY c.name
-            OFFSET :offset
-            LIMIT :limit
-        SQL;
-
-        $statement = $this->prepare($sql);
-        $statement->execute([
-            ':topic_id' => $topic_id,
-            ':offset' => $pagination_offset,
-            ':limit' => $pagination_limit,
         ]);
         return $statement->fetchAll();
     }
@@ -194,35 +90,6 @@ class Collection extends \Minz\DatabaseModel
             ':user_id' => $user_id,
         ]);
         return $statement->fetchAll();
-    }
-
-    /**
-     * Count the public collections attached to the given topic.
-     *
-     * @param string $topic_id
-     *
-     * @return integer
-     */
-    public function countPublicByTopicId($topic_id)
-    {
-        $sql = <<<'SQL'
-            SELECT COUNT(DISTINCT c.id)
-            FROM collections c, collections_to_topics ct, links_to_collections lc, links l
-
-            WHERE c.id = ct.collection_id
-            AND c.id = lc.collection_id
-            AND l.id = lc.link_id
-
-            AND c.is_public = true
-            AND l.is_hidden = false
-            AND ct.topic_id = :topic_id
-        SQL;
-
-        $statement = $this->prepare($sql);
-        $statement->execute([
-            ':topic_id' => $topic_id,
-        ]);
-        return intval($statement->fetchColumn());
     }
 
     /**
@@ -331,43 +198,6 @@ class Collection extends \Minz\DatabaseModel
     }
 
     /**
-     * Return the list of feeds of the given user and with the given feed URLs.
-     *
-     * @param string $user_id
-     * @param string[] $feed_urls
-     *
-     * @return array
-     */
-    public function listByUserIdAndFeedUrlsWithNumberLinks($user_id, $feed_urls)
-    {
-        if (!$feed_urls) {
-            return [];
-        }
-
-        $urls_as_question_marks = array_fill(0, count($feed_urls), '?');
-        $urls_where_statement = implode(', ', $urls_as_question_marks);
-
-        $sql = <<<SQL
-            SELECT c.*, COUNT(lc.id) AS number_links
-            FROM collections c
-
-            LEFT JOIN links_to_collections lc
-            ON lc.collection_id = c.id
-
-            WHERE c.user_id = ?
-            AND c.type = 'feed'
-            AND c.feed_url IN ({$urls_where_statement})
-            AND c.is_public = true
-
-            GROUP BY c.id
-        SQL;
-
-        $statement = $this->prepare($sql);
-        $statement->execute(array_merge([$user_id], $feed_urls));
-        return $statement->fetchAll();
-    }
-
-    /**
      * Return the followed collections of the given user and in the given group.
      *
      * If group id is null, it returns the collections in no groups.
@@ -432,213 +262,5 @@ class Collection extends \Minz\DatabaseModel
             'user_id' => $user_id,
         ]);
         return count($matching_rows) === count($collection_ids);
-    }
-
-    /**
-     * Return the list of ids indexed by names for the given user.
-     *
-     * @param string $user_id
-     *
-     * @return array
-     */
-    public function listNamesToIdsByUserId($user_id)
-    {
-        $sql = <<<SQL
-            SELECT id, name FROM collections
-            WHERE user_id = :user_id
-        SQL;
-
-        $statement = $this->prepare($sql);
-        $statement->execute([
-            ':user_id' => $user_id,
-        ]);
-
-        $ids_by_names = [];
-        foreach ($statement->fetchAll() as $row) {
-            $ids_by_names[$row['name']] = $row['id'];
-        }
-        return $ids_by_names;
-    }
-
-    /**
-     * Return the list of ids indexed by feed urls for the given user.
-     *
-     * @param string $user_id
-     *
-     * @return array
-     */
-    public function listFeedUrlsToIdsByUserId($user_id)
-    {
-        $sql = <<<SQL
-            SELECT id, feed_url FROM collections
-            WHERE user_id = :user_id
-            AND type = 'feed'
-        SQL;
-
-        $statement = $this->prepare($sql);
-        $statement->execute([
-            ':user_id' => $user_id,
-        ]);
-
-        $ids_by_feed_urls = [];
-        foreach ($statement->fetchAll() as $row) {
-            $ids_by_feed_urls[$row['feed_url']] = $row['id'];
-        }
-        return $ids_by_feed_urls;
-    }
-
-    /**
-     * List (randomly) active feeds to be fetched.
-     *
-     * An active feed is a feed followed by at least one active user (i.e.
-     * account has been validated)
-     *
-     * @param \DateTime $before
-     * @param integer $limit
-     *
-     * @return array
-     */
-    public function listActiveFeedsToFetch($before, $limit)
-    {
-        $sql = <<<SQL
-            SELECT c.*
-            FROM collections c, followed_collections fc, users u
-
-            WHERE c.type = 'feed'
-            AND (
-                c.feed_fetched_at <= :before
-                OR c.feed_fetched_at IS NULL
-            )
-
-            AND c.id = fc.collection_id
-            AND u.id = fc.user_id
-
-            -- We prioritize feeds followed by active users. We ignore for now
-            -- expired subscriptions, but it could be checked too.
-            AND u.validated_at IS NOT NULL
-
-            ORDER BY random()
-            LIMIT :limit
-        SQL;
-
-        $statement = $this->prepare($sql);
-        $statement->execute([
-            ':before' => $before->format(\Minz\Model::DATETIME_FORMAT),
-            ':limit' => $limit,
-        ]);
-        return $statement->fetchAll();
-    }
-
-    /**
-     * List feeds that haven't been fetched for the longest time.
-     *
-     * @param \DateTime $before
-     * @param integer $limit
-     *
-     * @return array
-     */
-    public function listOldestFeedsToFetch($before, $limit)
-    {
-        $sql = <<<SQL
-            SELECT c.*
-            FROM collections c
-
-            WHERE c.type = 'feed'
-            AND (
-                c.feed_fetched_at <= :before
-                OR c.feed_fetched_at IS NULL
-            )
-
-            ORDER BY feed_fetched_at NULLS FIRST
-            LIMIT :limit
-        SQL;
-
-        $statement = $this->prepare($sql);
-        $statement->execute([
-            ':before' => $before->format(\Minz\Model::DATETIME_FORMAT),
-            ':limit' => $limit,
-        ]);
-        return $statement->fetchAll();
-    }
-
-    /**
-     * Delete not followed collections older than the given date for the given
-     * user.
-     *
-     * @param string $user_id
-     * @param \DateTime $date
-     *
-     * @return boolean True on success
-     */
-    public function deleteUnfollowedOlderThan($user_id, $date)
-    {
-        $sql = <<<SQL
-            DELETE FROM collections
-
-            USING collections AS c
-
-            LEFT JOIN followed_collections AS fc
-            ON c.id = fc.collection_id
-
-            WHERE collections.id = c.id
-            AND c.user_id = :user_id
-            AND c.created_at < :date
-            AND fc.collection_id IS NULL;
-        SQL;
-
-        $statement = $this->prepare($sql);
-        return $statement->execute([
-            ':user_id' => $user_id,
-            ':date' => $date->format(\Minz\Model::DATETIME_FORMAT),
-        ]);
-    }
-
-    /**
-     * Lock a collection
-     *
-     * @param string $collection_id
-     *
-     * @return boolean True if the lock is successful, false otherwise
-     */
-    public function lock($collection_id)
-    {
-        $sql = <<<SQL
-            UPDATE collections
-            SET locked_at = :locked_at
-            WHERE id = :collection_id
-            AND (locked_at IS NULL OR locked_at <= :lock_timeout)
-        SQL;
-
-        $now = \Minz\Time::now();
-        $lock_timeout = \Minz\Time::ago(1, 'hour');
-        $statement = $this->prepare($sql);
-        $statement->execute([
-            ':locked_at' => $now->format(\Minz\Model::DATETIME_FORMAT),
-            ':collection_id' => $collection_id,
-            ':lock_timeout' => $lock_timeout->format(\Minz\Model::DATETIME_FORMAT),
-        ]);
-        return $statement->rowCount() === 1;
-    }
-
-    /**
-     * Unlock a collection
-     *
-     * @param string $collection_id
-     *
-     * @return boolean True if the unlock is successful, false otherwise
-     */
-    public function unlock($collection_id)
-    {
-        $sql = <<<SQL
-            UPDATE collections
-            SET locked_at = null
-            WHERE id = :collection_id
-        SQL;
-
-        $statement = $this->prepare($sql);
-        $statement->execute([
-            ':collection_id' => $collection_id,
-        ]);
-        return $statement->rowCount() === 1;
     }
 }
